# DF/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user/schema": {
      "get": {
        "description": "Retrive the datatype of the user table",
        "operationId": "userController_findSchema",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "schema validation",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/user/get": {
      "get": {
        "description": "Filter all the records from the user table",
        "operationId": "userController_findAllmethod",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of records to fetch",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/userEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Filter the records",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/user/{id}": {
      "get": {
        "description": "Read only one records from the user table",
        "operationId": "userController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/userEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Fetch the only one record",
        "tags": [
          "ERD API"
        ]
      },
      "patch": {
        "description": "Update the record for the user table",
        "operationId": "userController_update",
        "parameters": [
          {
            "name": "xCdcaRole",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xCdcaUsername",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xCdcaRemarks",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xCdcaApprovalStatus",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateuserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/userEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update the record",
        "tags": [
          "ERD API"
        ]
      },
      "delete": {
        "description": "Delete the record for the user table",
        "operationId": "userController_remove",
        "parameters": [
          {
            "name": "xCdcaRole",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xCdcaUsername",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xCdcaRemarks",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xCdcaApprovalStatus",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/userEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete the record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/user": {
      "get": {
        "description": "Read all the records from the user table",
        "operationId": "userController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/userEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Read all the records",
        "tags": [
          "ERD API"
        ]
      },
      "post": {
        "description": "Create the record for the user table",
        "operationId": "userController_create",
        "parameters": [
          {
            "name": "xCdcaRole",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xCdcaUsername",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xCdcaRemarks",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xCdcaApprovalStatus",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xCdcaApprovalID",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateuserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create the record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/user/find/first": {
      "get": {
        "description": "Read first record from the user table",
        "operationId": "userController_findFirst",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Fetch the first record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/user/find/last": {
      "get": {
        "description": "Read last record from the user table",
        "operationId": "userController_findLast",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Fetch the last record",
        "tags": [
          "ERD API"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nest API",
    "description": "The Nestjs API documentation",
    "version": "0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://tgadev2.toruslowcode.com/ct001/tgw/tgw004/v1/api",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "userEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": "number"
          },
          "name": {
            "type": "string",
            "example": "string"
          },
          "onlydate": {
            "format": "date-time",
            "type": "string",
            "example": "date"
          },
          "dateandtime": {
            "format": "date-time",
            "type": "string",
            "example": "datetime"
          },
          "trs_created_date": {
            "format": "date-time",
            "type": "string",
            "example": "datetime"
          },
          "trs_created_by": {
            "type": "string",
            "example": "string"
          },
          "trs_modified_date": {
            "format": "date-time",
            "type": "string",
            "example": "datetime"
          },
          "trs_modified_by": {
            "type": "string",
            "example": "string"
          },
          "trs_process_id": {
            "type": "string",
            "example": "string"
          },
          "trs_access_profile": {
            "type": "string",
            "example": "string"
          },
          "trs_org_grp_code": {
            "type": "string",
            "example": "string"
          },
          "trs_org_code": {
            "type": "string",
            "example": "string"
          },
          "trs_role_grp_code": {
            "type": "string",
            "example": "string"
          },
          "trs_role_code": {
            "type": "string",
            "example": "string"
          },
          "trs_ps_grp_code": {
            "type": "string",
            "example": "string"
          },
          "trs_ps_code": {
            "type": "string",
            "example": "string"
          },
          "trs_sub_org_grp_code": {
            "type": "string",
            "example": "string"
          },
          "trs_sub_org_code": {
            "type": "string",
            "example": "string"
          },
          "trs_locked_by": {
            "type": "string",
            "example": "string"
          },
          "trs_locked_time": {
            "format": "date-time",
            "type": "string",
            "example": "datetime"
          },
          "trs_tenant_id": {
            "type": "string",
            "example": "string"
          },
          "trs_app_code": {
            "type": "string",
            "example": "string"
          },
          "trs_product_code": {
            "type": "string",
            "example": "string"
          },
          "trs_event_process_status": {
            "type": "string",
            "example": "string"
          },
          "trs_event_status": {
            "type": "string",
            "example": "string"
          },
          "trs_token_id": {
            "type": "string",
            "example": "string"
          },
          "trs_version": {
            "type": "string",
            "example": "string"
          },
          "trs_prev_process_code": {
            "type": "string",
            "example": "string"
          },
          "trs_prev_status": {
            "type": "string",
            "example": "string"
          },
          "trs_prev_process_status": {
            "type": "string",
            "example": "string"
          },
          "trs_process_code": {
            "type": "string",
            "example": "string"
          },
          "trs_status": {
            "type": "string",
            "example": "string"
          },
          "trs_process_status": {
            "type": "string",
            "example": "string"
          },
          "trs_next_process_code": {
            "type": "string",
            "example": "string"
          },
          "trs_next_status": {
            "type": "string",
            "example": "string"
          },
          "trs_next_process_status": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "CreateuserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "onlydate": {
            "type": "string",
            "format": "date-time"
          },
          "dateandtime": {
            "type": "string",
            "format": "date-time"
          },
          "trs_created_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          },
          "trs_sub_org_grp_code": {
            "type": "string"
          },
          "trs_sub_org_code": {
            "type": "string"
          },
          "trs_locked_by": {
            "type": "string"
          },
          "trs_locked_time": {
            "type": "string",
            "format": "date-time"
          },
          "trs_tenant_id": {
            "type": "string"
          },
          "trs_app_code": {
            "type": "string"
          },
          "trs_product_code": {
            "type": "string"
          },
          "trs_event_process_status": {
            "type": "string"
          },
          "trs_event_status": {
            "type": "string"
          },
          "trs_token_id": {
            "type": "string"
          },
          "trs_version": {
            "type": "string"
          },
          "trs_prev_process_code": {
            "type": "string"
          },
          "trs_prev_status": {
            "type": "string"
          },
          "trs_prev_process_status": {
            "type": "string"
          },
          "trs_process_code": {
            "type": "string"
          },
          "trs_status": {
            "type": "string"
          },
          "trs_process_status": {
            "type": "string"
          },
          "trs_next_process_code": {
            "type": "string"
          },
          "trs_next_status": {
            "type": "string"
          },
          "trs_next_process_status": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "UpdateuserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "onlydate": {
            "type": "string",
            "format": "date-time"
          },
          "dateandtime": {
            "type": "string",
            "format": "date-time"
          },
          "trs_created_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          },
          "trs_sub_org_grp_code": {
            "type": "string"
          },
          "trs_sub_org_code": {
            "type": "string"
          },
          "trs_locked_by": {
            "type": "string"
          },
          "trs_locked_time": {
            "type": "string",
            "format": "date-time"
          },
          "trs_tenant_id": {
            "type": "string"
          },
          "trs_app_code": {
            "type": "string"
          },
          "trs_product_code": {
            "type": "string"
          },
          "trs_event_process_status": {
            "type": "string"
          },
          "trs_event_status": {
            "type": "string"
          },
          "trs_token_id": {
            "type": "string"
          },
          "trs_version": {
            "type": "string"
          },
          "trs_prev_process_code": {
            "type": "string"
          },
          "trs_prev_status": {
            "type": "string"
          },
          "trs_prev_process_status": {
            "type": "string"
          },
          "trs_process_code": {
            "type": "string"
          },
          "trs_status": {
            "type": "string"
          },
          "trs_process_status": {
            "type": "string"
          },
          "trs_next_process_code": {
            "type": "string"
          },
          "trs_next_status": {
            "type": "string"
          },
          "trs_next_process_status": {
            "type": "string"
          }
        }
      },
...
    }
  }
}